deploy/requirements.txt
sqlalchemy
pandas
requests
beautifulsoup4
selenium
webdriver-manager
psycopg2-binary
pymysql
Flask
python-dotenv
lxml